Revert "useDebounceEvent.stories.tsx - temp fix for the hook issue"
This reverts commit 61813c65dd4a1c044997b959a719a2d1d4015262.
